# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Center API",
    "summary": "Документ-центр ИП «Северный Контур»: проекты, документы, версии и аудит",
    "description": "\nВнутренний REST API ИП «Северный Контур» для управления документами по проектам.\n\n**Базовый путь публичного контракта:** `/api/v1`. Версия контракта (`v1`)\nне совпадает с версией приложения, указанной в заголовке документации.\n\n### Аутентификация\nПолучите токен операцией `POST /api/v1/auth/login` и передавайте его в заголовке\n`Authorization: Bearer <token>`. В Swagger UI нажмите **Authorize** и вставьте токен.\n\n### Роли\n- **admin** — полный доступ ко всем проектам, пользователям и журналу действий;\n- **manager** — создаёт проекты и выдаёт доступ участникам своих проектов;\n- **worker** — работает с документами в рамках выданного доступа `editor`;\n- **viewer** — только чтение в рамках выданного доступа.\n\n### Ошибки\nВсе ошибки возвращаются в формате RFC 9457 Problem Details\n(`application/problem+json`) с полями `type`, `title`, `status`, `detail`,\n`instance`; ошибки валидации дополнительно содержат массив `errors`.\n\n### Коллекции\nКоллекции поддерживают параметры `offset` (≥ 0) и `limit` (1…100, по умолчанию 20)\nи возвращают конверт `items`, `total`, `offset`, `limit`.\n",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Получить список пользователей",
        "description": "Вернуть учётные записи пользователей.\n\nОперация доступна только администратору.",
        "operationId": "list_users",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Отбор по роли пользователя",
              "title": "Role"
            },
            "description": "Отбор по роли пользователя"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение от начала коллекции",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение от начала коллекции"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы, максимум 100",
              "default": 20,
              "title": "Limit"
            },
            "description": "Размер страницы, максимум 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Страница учётных записей",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserRead_"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Создать пользователя",
        "description": "Создать учётную запись пользователя.\n\nОперация доступна только администратору. Самостоятельная регистрация\nв системе не предусмотрена.",
        "operationId": "create_user",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Созданная учётная запись",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Изменить пользователя",
        "description": "Изменить роль или состояние учётной записи.\n\nНеуказанные поля не изменяются. Администратор не может деактивировать\nсобственную учётную запись.",
        "operationId": "update_user",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Идентификатор пользователя",
              "title": "User Id"
            },
            "description": "Идентификатор пользователя"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Изменённая учётная запись",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Путь к полю запроса",
            "examples": [
              "title"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Причина отклонения значения",
            "examples": [
              "String should have at least 3 characters"
            ]
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "description": "Описание одной ошибки валидации поля."
      },
...
      "Page_UserRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserRead"
            },
            "type": "array",
            "title": "Items",
            "description": "Записи текущей страницы"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Общее количество записей, доступных запросу"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Смещение текущей страницы от начала коллекции"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Запрошенный размер страницы"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page[UserRead]",
        "description": "Страница коллекции с метаданными пагинации: items, total, offset, limit"
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный URI типа ошибки"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Краткое название типа ошибки, не зависящее от экземпляра"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "description": "HTTP-код ответа"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Описание конкретного случая возникновения ошибки"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "Путь запроса, при обработке которого возникла ошибка"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ProblemDetails",
        "description": "Тело ответа с ошибкой по RFC 9457.",
        "examples": [
          {
            "detail": "Project 42 does not exist or is unavailable",
            "instance": "/api/v1/projects/42",
            "status": 404,
            "title": "Project not found",
            "type": "https://docs.local/problems/project-not-found"
          }
        ]
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Адрес электронной почты",
            "examples": [
              "user@example.com"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Пароль: не менее 8 символов, содержит букву и цифру",
            "examples": [
              "Passw0rd123"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Назначаемая роль",
            "default": "viewer"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserCreate",
        "description": "Данные для создания учётной записи."
      },
...
      "UserRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Идентификатор пользователя"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Адрес электронной почты"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Роль пользователя"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Признак активности учётной записи"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Момент создания учётной записи",
            "examples": [
              "2026-09-09T08:15:30Z"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserRead",
        "description": "Карточка пользователя.\n\nХэш пароля и иные внутренние атрибуты в контракт не входят."
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "manager",
          "worker",
          "viewer"
        ],
        "title": "UserRole",
        "description": "Роль пользователя: admin, manager, worker или viewer."
      },
      "UserUpdate": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              },
              {
                "type": "null"
              }
            ],
            "description": "Новая роль пользователя"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "Признак активности учётной записи"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Изменяемые атрибуты учётной записи."
      },
      "ValidationProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный URI типа ошибки"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Краткое название типа ошибки, не зависящее от экземпляра"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "description": "HTTP-код ответа"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Описание конкретного случая возникновения ошибки"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "Путь запроса, при обработке которого возникла ошибка"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors",
            "description": "Перечень полей, не прошедших валидацию"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ValidationProblemDetails",
        "description": "Тело ответа для ошибок валидации запроса (HTTP 422).\n\nРасширяет базовую структуру членом ``errors`` согласно STD-ERR-05.",
        "examples": [
          {
            "detail": "Request body does not match the expected schema",
            "errors": [
              {
                "field": "title",
                "message": "String should have at least 3 characters"
              }
            ],
            "instance": "/api/v1/projects",
            "status": 422,
            "title": "Request validation failed",
            "type": "https://docs.local/problems/validation-error"
          }
        ]
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}